+++ core/src/components/cat-checkbox/cat-checkbox.tsx
@@ -1,8 +1,15 @@
-import { Component, Event, EventEmitter, h, Prop } from '@stencil/core';
+import { Component, Event, EventEmitter, h, Method, Prop } from '@stencil/core';
 import log from 'loglevel';
 
 let nextUniqueId = 0;
 
+/**
+ * Checkboxes are used to let a user choose one or more options from a limited
+ * number of options.
+ *
+ * @part checkbox - The checkbox element.
+ * @part label - The label content.
+ */
 @Component({
   tag: 'cat-checkbox',
   styleUrls: ['cat-checkbox.scss'],
@@ -10,7 +17,7 @@ let nextUniqueId = 0;
 })
 export class CatCheckbox {
   private readonly id = `cat-checkbox-${nextUniqueId++}`;
-  private inputRef?: HTMLInputElement;
+  private input!: HTMLInputElement;
 
   /**
    * Checked state of the checkbox
@@ -53,13 +60,23 @@ export class CatCheckbox {
   @Prop() value?: string;
 
   /**
-   * Emitted when the checked status of the checkbox is changed
+   * Emitted when the checked status of the checkbox is changed.
    */
-  @Event() checkboxChange!: EventEmitter;
+  @Event() catChange!: EventEmitter;
+
+  /**
+   * Emitted when the checkbox received focus.
+   */
+  @Event() catFocus!: EventEmitter<FocusEvent>;
+
+  /**
+   * Emitted when the checkbox loses focus.
+   */
+  @Event() catBlur!: EventEmitter<FocusEvent>;
 
   componentDidLoad() {
-    if (this.inputRef && this.indeterminate) {
-      this.inputRef.indeterminate = true;
+    if (this.input && this.indeterminate) {
+      this.input.indeterminate = true;
     }
   }
 
@@ -69,11 +86,22 @@ export class CatCheckbox {
     }
   }
 
+  /**
+   * Sets focus on the checkbox. Use this method instead of `checkbox.focus()`.
+   *
+   * @param options An optional object providing options to control aspects of
+   * the focusing process.
+   */
+  @Method()
+  async setFocus(options?: FocusOptions): Promise<void> {
+    this.input.focus(options);
+  }
+
   render() {
     return (
       <label htmlFor={this.id} class={{ 'is-hidden': this.labelHidden, 'is-disabled': this.disabled }}>
         <input
-          ref={el => (this.inputRef = el)}
+          ref={el => (this.input = el as HTMLInputElement)}
           id={this.id}
           type="checkbox"
           name={this.name}
@@ -81,9 +109,11 @@ export class CatCheckbox {
           checked={this.checked}
           required={this.required}
           disabled={this.disabled}
-          onInput={this.handleChange.bind(this)}
+          onInput={this.onInput.bind(this)}
+          onFocus={this.onFocus.bind(this)}
+          onBlur={this.onBlur.bind(this)}
         />
-        <span class="box" aria-hidden="true">
+        <span class="box" aria-hidden="true" part="checkbox">
           <svg class="check" viewBox="0 0 12 10">
             <polyline points="1.5 6 4.5 9 10.5 1"></polyline>
           </svg>
@@ -91,12 +121,22 @@ export class CatCheckbox {
             <polyline points="1.5 5 10.5 5"></polyline>
           </svg>
         </span>
-        <span class="label">{this.label}</span>
+        <span class="label" part="label">
+          {this.label}
+        </span>
       </label>
     );
   }
 
-  private handleChange(event: Event) {
-    this.checkboxChange.emit(event);
+  private onInput(event: Event) {
+    this.catChange.emit(event);
+  }
+
+  private onFocus(event: FocusEvent) {
+    this.catFocus.emit(event);
+  }
+
+  private onBlur(event: FocusEvent) {
+    this.catBlur.emit(event);
   }
 }