+++ core/src/components/cat-textarea/cat-textarea.e2e.ts
@@ -1,0 +1,11 @@
+import { newE2EPage } from '@stencil/core/testing';
+
+describe('cat-textarea', () => {
+  it('renders', async () => {
+    const page = await newE2EPage();
+    await page.setContent('<cat-textarea label="Label"></cat-textarea>');
+
+    const element = await page.find('cat-textarea');
+    expect(element).toHaveClass('hydrated');
+  });
+});