+++ react/src/components/stencil-generated/index.ts
@@ -10,6 +10,7 @@ import type { JSX } from '@haiilo/catalyst';
 export const CatAlert = /*@__PURE__*/createReactComponent<JSX.CatAlert, HTMLCatAlertElement>('cat-alert');
 export const CatBadge = /*@__PURE__*/createReactComponent<JSX.CatBadge, HTMLCatBadgeElement>('cat-badge');
 export const CatButton = /*@__PURE__*/createReactComponent<JSX.CatButton, HTMLCatButtonElement>('cat-button');
+export const CatCard = /*@__PURE__*/createReactComponent<JSX.CatCard, HTMLCatCardElement>('cat-card');
 export const CatCheckbox = /*@__PURE__*/createReactComponent<JSX.CatCheckbox, HTMLCatCheckboxElement>('cat-checkbox');
 export const CatIcon = /*@__PURE__*/createReactComponent<JSX.CatIcon, HTMLCatIconElement>('cat-icon');
 export const CatInput = /*@__PURE__*/createReactComponent<JSX.CatInput, HTMLCatInputElement>('cat-input');
@@ -18,5 +19,7 @@ export const CatRadio = /*@__PURE__*/createReactComponent<JSX.CatRadio, HTMLCatR
 export const CatScrollable = /*@__PURE__*/createReactComponent<JSX.CatScrollable, HTMLCatScrollableElement>('cat-scrollable');
 export const CatSkeleton = /*@__PURE__*/createReactComponent<JSX.CatSkeleton, HTMLCatSkeletonElement>('cat-skeleton');
 export const CatSpinner = /*@__PURE__*/createReactComponent<JSX.CatSpinner, HTMLCatSpinnerElement>('cat-spinner');
+export const CatToast = /*@__PURE__*/createReactComponent<JSX.CatToast, HTMLCatToastElement>('cat-toast');
+export const CatTextarea = /*@__PURE__*/createReactComponent<JSX.CatTextarea, HTMLCatTextareaElement>('cat-textarea');
 export const CatToggle = /*@__PURE__*/createReactComponent<JSX.CatToggle, HTMLCatToggleElement>('cat-toggle');
 export const CatTooltip = /*@__PURE__*/createReactComponent<JSX.CatTooltip, HTMLCatTooltipElement>('cat-tooltip');