+++ tokens/src/color/ui.js
@@ -7,7 +7,6 @@ module.exports = {
         interaction: { value: '{color.base.brand.400.value}' },
         input: { value: '{color.base.neutral.0.value}' },
         inputDisabled: { value: '{color.base.neutral.100.value}' },
-        notification: { value: '{color.base.red.400.value}' },
         skeleton: { value: '{color.base.neutral.200.value}' },
         skeletonHighlight: { value: '{color.base.neutral.300.value}' },
         tooltip: { value: '{color.base.neutral.0.value}' }
@@ -16,7 +15,7 @@ module.exports = {
         light: { value: '{color.base.neutral.100.value}' },
         default: { value: '{color.base.neutral.200.value}' },
         dark: { value: '{color.base.neutral.300.value}' },
-        focus: { value: '#0071ff' },
+        focus: { value: '#0071ff', attributes: { cssProp: "border-color-focus" }  },
         input: { value: '{color.base.neutral.300.value}' }
       },
       font: {
@@ -25,7 +24,6 @@ module.exports = {
         mono: { value: '{color.base.neutral.900.value}', attributes: { cssProp: "font-color-mono" } },
         muted: { value: '{color.base.neutral.400.value}', attributes: { cssProp: "font-color-muted" } },
         quote: { value: '{color.ui.font.body.value}' },
-        quoteSource: { value: '{color.ui.font.muted.value}' },
         tooltip: { value: '{color.base.neutral.900.value}' }
       }
     }